Refactors handlers into individual files
Change-type: patch
