+++ lib/action-library/handlers/action-oauth-authorize.js
@@ -1,0 +1,42 @@
+/*
+ * Copyright (C) Balena.io - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited.
+ * Proprietary and confidential.
+ */
+
+const sync = require('../../sync')
+const environment = require('../../environment')
+const syncContext = require('./sync-context')
+
+const handler = async (session, context, card, request) => {
+	const syncContextInstance = syncContext.fromWorkerContext(request.arguments.provider,
+		context, request.context, session)
+
+	const credentials = await sync.authorize(
+		request.arguments.provider,
+		environment.integration[request.arguments.provider],
+		syncContextInstance,
+		{
+			code: request.arguments.code,
+			origin: request.arguments.origin
+		}
+	)
+
+	const user = await sync.whoami(
+		request.arguments.provider,
+		syncContextInstance,
+		{
+			credentials,
+			slug: request.arguments.slug
+		}
+	)
+
+	return {
+		user,
+		credentials
+	}
+}
+
+module.exports = {
+	handler
+}
