+++ lib/action-library/handlers/sync-context.js
@@ -6,8 +6,8 @@
 
 const _ = require('lodash')
 const jsonpatch = require('fast-json-patch')
-const logger = require('../logger').getLogger(__filename)
-const assert = require('../assert')
+const logger = require('../../logger').getLogger(__filename)
+const assert = require('../../assert')
 
 /*
  * Deal with some username differences that we can't
